requirements.txt
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
blinker==1.3
Brlapi==0.6.4
catfish==1.4.2
certifi==2018.8.24
cffi==1.5.2
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.2.3
defer==1.0.6
devscripts===2.16.2ubuntu3
dirspec==13.10
Django==2.0.6
django-rest-swagger==2.2.0
djangorestframework==3.8.2
dnspython==1.15.0
ecdsa==0.13
Electrum==3.0.5
httplib2==0.9.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
jsonrpclib-pelix==0.3.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==1.0
menulibre==2.1.3
mugshot==0.3.1
oauthlib==1.0.3
oneconf==0.3.9
openapi-codec==1.3.2
pbkdf2==1.3
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
ply==3.7
protobuf==3.5.1
psutil==3.4.2
ptyprocess==0.5
pyaes==1.6.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyICU==1.9.2
PyJWT==1.3.0
PySocks==1.6.8
python-apt==1.1.0b1
python-debian==0.1.27
pytz==2018.4
pyxdg==0.25
qrcode==5.3
reportlab==3.3.0
requests==2.19.1
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
simplejson==3.16.0
six==1.10.0
system-service==0.3
thin-client-config-agent==0.8
ubuntu-drivers-common==0.0.0
ufw==0.35
uritemplate==3.0.0
urllib3==1.23
virtkey==0.63.0
xkit==0.0.0
